move discovery of Guice modules required by  client plugins to NdBenchClientModule
